Abstracting all of the 2 pages of keybindings
I think it's best to encourage users to get to code writing regardless of their knowledge of keybindings, and as they code they will learn keybindings naturally. Also the mouse is a direct innovation to utilizing purely keyboards, it makes computer far more accessible, and its nice to encourage that.